requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
pydantic==2.11.9
orjson==3.11.3
numpy==2.3.3
pandas==2.3.2
pyarrow==21.0.0
boto3==1.40.40
python-dotenv==1.1.1
yfinance==0.2.66
requests>=2.31.0
openai==1.40.0
xai-sdk==1.3.1
s3fs==2024.12.0
